refactor: extract fallback executor + add model strategy policy
