+++ src/lib/policy/defaults/rules/gpt.ts
@@ -1,5 +1,17 @@
-import type { ModelRoutingRule, ProviderId } from "../../schema.js";
+import type { ModelRoutingRule, ProviderId, UpstreamMode } from "../../schema.js";
 import { GPT_FREE_BLOCKED_MODEL_PATTERN, PAID_PLAN_WEIGHTS } from "./plans.js";
+
+const GPT_DEFAULT_STRATEGIES: readonly UpstreamMode[] = [
+  "openai_responses",
+  "responses",
+  "openai_chat_completions",
+  "chat_completions",
+];
+
+const GPT_OSS_STRATEGIES: readonly UpstreamMode[] = [
+  "ollama_chat",
+  "chat_completions",
+];
 
 export const DEFAULT_GPT_PROVIDER_ORDER: readonly ProviderId[] = [
   "openai",
@@ -24,6 +36,7 @@ export function createGptRoutingRules(): readonly ModelRoutingRule[] {
     {
       modelPattern: /^gpt-oss/,
       preferredProviders: GPT_OSS_PROVIDER_ORDER,
+      preferredStrategies: GPT_OSS_STRATEGIES,
       accountOrdering: { kind: "prefer_free" },
     },
     // GPT blocked models (free tier blocked, requires paid plan)
@@ -32,6 +45,7 @@ export function createGptRoutingRules(): readonly ModelRoutingRule[] {
       requiresPaidPlan: true,
       preferredProviders: DEFAULT_GPT_PROVIDER_ORDER,
       excludedProviders: GPT_EXCLUDED_PROVIDERS,
+      preferredStrategies: GPT_DEFAULT_STRATEGIES,
       accountOrdering: { kind: "custom_weight", weights: PAID_PLAN_WEIGHTS },
     },
     // GPT 6+ (requires paid plan)
@@ -40,6 +54,7 @@ export function createGptRoutingRules(): readonly ModelRoutingRule[] {
       requiresPaidPlan: true,
       preferredProviders: DEFAULT_GPT_PROVIDER_ORDER,
       excludedProviders: GPT_EXCLUDED_PROVIDERS,
+      preferredStrategies: GPT_DEFAULT_STRATEGIES,
       accountOrdering: { kind: "custom_weight", weights: PAID_PLAN_WEIGHTS },
     },
     // GPT general routing (catch-all for remaining gpt-* models)
@@ -47,6 +62,7 @@ export function createGptRoutingRules(): readonly ModelRoutingRule[] {
       modelPattern: /^gpt-/,
       preferredProviders: DEFAULT_GPT_PROVIDER_ORDER,
       excludedProviders: GPT_EXCLUDED_PROVIDERS,
+      preferredStrategies: GPT_DEFAULT_STRATEGIES,
       accountOrdering: { kind: "prefer_free" },
     },
   ];
