+++ src/lib/provider-strategy/fallback/index.ts
@@ -1,43 +1,3 @@
-export {
-  executeProviderRoutingPlan,
-  executeProviderFallback,
-  inspectProviderAvailability,
-} from "./legacy.js";
-
-export type {
-  ErrorClassification,
-  ErrorClassificationResult,
-} from "./error-classifier.js";
-
-export {
-  classifyAuthError,
-  classifyModelNotSupported,
-  shouldCooldownCredentialOnAuthFailure,
-  shouldPermanentlyDisableCredential,
-  shouldRetrySameCredentialForServerError,
-  PERMANENT_DISABLE_COOLDOWN_MS,
-} from "./error-classifier.js";
-
-export {
-  providerAccountsForRequest,
-  providerAccountsForRequestWithPolicy,
-  reorderAccountsForLatency,
-  reorderCandidatesForAffinities,
-  reorderCandidatesForAffinity,
-  gptModelRequiresPaidPlan,
-} from "./credential-selector.js";
-
-export type { PreferredAffinity } from "./credential-selector.js";
-
-export type { FallbackCandidate, FallbackDeps, FallbackKeyPool } from "./types.js";
-
-export {
-  clampRouteQuality,
-  createAccumulator,
-  emptyResult,
-  successResult,
-} from "./types.js";
-
-export { buildFallbackCandidates } from "./orchestrator.js";
-
-export type { BuildCandidatesResult } from "./orchestrator.js";
+// Compatibility layer: the new execution+candidate code lives under
+// `provider-strategy/routing/`. The `fallback/` path remains for external imports.
+export * from "../routing/index.js";
