+++ src/lib/provider-strategy/fallback/legacy.ts
@@ -1,1185 +1,3 @@
-import type { FastifyReply } from "fastify";
-
-import type { AccountHealthStore } from "../../db/account-health-store.js";
-import type { EventStore } from "../../db/event-store.js";
-import type { ProviderCredential } from "../../key-pool.js";
-import type { PolicyEngine } from "../../policy/index.js";
-import type { IPromptAffinityStore } from "../../prompt-affinity-store.js";
-import type { ProviderRoutePheromoneStore } from "../../provider-route-pheromone-store.js";
-import type { RequestLogStore } from "../../request-log-store.js";
-import type { QuotaMonitor } from "../../quota-monitor.js";
-import {
-  buildUpstreamHeadersForCredential,
-  classifyRateLimitKind,
-  detectOllamaLimitKind,
-  extractRateLimitCooldownMs,
-  isRateLimitResponse,
-} from "../../proxy.js";
-import {
-  responsesEventStreamToErrorPayload,
-} from "../../responses-compat.js";
-import type { ProviderRoute } from "../../provider-routing.js";
-import { fetchWithResponseTimeout } from "../../http/index.js";
-import {
-  responseIndicatesQuotaError,
-  summarizeUpstreamError,
-  toErrorMessage,
-} from "../../errors/index.js";
-import { getTelemetry } from "../../telemetry/otel.js";
-import { selectRemoteProviderStrategyForRoute } from "../registry.js";
-import {
-  buildCodexResponsesImagesBody,
-  buildFactory4xxDiagnostics,
-  extractImagesFromCodexEventStream,
-  extractImagesFromCodexResponse,
-  joinUrl,
-  responseLooksLikeEventStream,
-  sleep,
-  transientRetryDelayMs,
-  recordAttempt,
-  updateFailedAttemptDiagnostics,
-  updateUsageCountsFromResponse,
-  type BuildPayloadResult,
-  type ProviderAttemptContext,
-  type ProviderAvailabilitySummary,
-  type ProviderFallbackExecutionResult,
-  type ProviderStrategy,
-  type StrategyRequestContext,
-} from "../shared.js";
-import {
-  PERMANENT_DISABLE_COOLDOWN_MS,
-  shouldCooldownCredentialOnAuthFailure,
-  shouldPermanentlyDisableCredential,
-  shouldRetrySameCredentialForServerError,
-} from "./error-classifier.js";
-import { requestyModelProvider } from "../../model-family.js";
-import { buildFallbackCandidates } from "./orchestrator.js";
-
-function isRecord(value: unknown): value is Record<string, unknown> {
-  return typeof value === "object" && value !== null;
-}
-
-function asString(value: unknown): string | undefined {
-  return typeof value === "string" ? value : undefined;
-}
-
-function normalizeReasoningEffortForOllamaCloud(payload: Record<string, unknown>): Record<string, unknown> {
-  const reasoning = isRecord(payload["reasoning"]) ? payload["reasoning"] : null;
-  const effort = asString(reasoning?.["effort"]) ?? asString(payload["reasoning_effort"]) ?? asString(payload["reasoningEffort"]);
-
-  if (!effort || effort.toLowerCase() !== "xhigh") {
-    return payload;
-  }
-
-  const result = { ...payload };
-
-  if (reasoning && asString(reasoning["effort"])?.toLowerCase() === "xhigh") {
-    result["reasoning"] = { ...reasoning, effort: "high" };
-  }
-
-  if (asString(payload["reasoning_effort"])?.toLowerCase() === "xhigh") {
-    result["reasoning_effort"] = "high";
-  }
-
-  if (asString(payload["reasoningEffort"])?.toLowerCase() === "xhigh") {
-    result["reasoningEffort"] = "high";
-  }
-
-  return result;
-}
-import { clampRouteQuality, createAccumulator, emptyResult, type FallbackDeps } from "./types.js";
-
-function shouldUseOpenAiCodexHeaderProfile(
-  providerId: string,
-  account: ProviderCredential,
-  openaiProviderId: string,
-): boolean {
-  return providerId === openaiProviderId && account.authType === "oauth_bearer";
-}
-
-const MAX_STICKY_TRANSPORT_FAILURE_CANDIDATES = 4;
-
-function requestyModelPrefix(model: string): string {
-  return requestyModelProvider(model);
-}
-
-function candidateMatchesAffinity(
-  candidate: { readonly providerId: string; readonly account: ProviderCredential },
-  affinity: { readonly providerId: string; readonly accountId: string } | undefined,
-): boolean {
-  return affinity !== undefined
-    && candidate.providerId === affinity.providerId
-    && candidate.account.accountId === affinity.accountId;
-}
-
-/**
- * Executes the provider routing plan: iterates over fallback candidates,
- * attempts each one, and handles success/failure/rate-limit outcomes.
- * Supports prompt-cache-key affinity for session stickiness and
- * hidden quota error detection (200 OK with error in stream body).
- */
-export async function executeProviderRoutingPlan(
-  strategy: ProviderStrategy,
-  reply: FastifyReply,
-  requestLogStore: RequestLogStore,
-  promptAffinityStore: IPromptAffinityStore,
-  providerRoutePheromoneStore: ProviderRoutePheromoneStore,
-  keyPool: {
-    getRequestOrder(providerId: string): Promise<ProviderCredential[]>;
-    markInFlight(credential: ProviderCredential): () => void;
-    markRateLimited(credential: ProviderCredential, retryAfterMs?: number): void;
-    isAccountExpired?(credential: ProviderCredential): boolean;
-    clearAccountCooldown?(providerId: string, accountId: string): void;
-    disableAccount?(providerId: string, accountId: string): void;
-  },
-  providerRoutes: readonly ProviderRoute[],
-  context: StrategyRequestContext,
-  payload: BuildPayloadResult,
-  promptCacheKey?: string,
-  refreshExpiredToken?: (credential: ProviderCredential) => Promise<ProviderCredential | null>,
-  policy?: PolicyEngine,
-  healthStore?: AccountHealthStore,
-  eventStore?: EventStore,
-  quotaMonitor?: QuotaMonitor,
-): Promise<ProviderFallbackExecutionResult> {
-  const accumulator = createAccumulator();
-
-  const deps: FallbackDeps = {
-    strategy, reply, requestLogStore, promptAffinityStore, providerRoutePheromoneStore,
-    keyPool, providerRoutes, context, payload, promptCacheKey, refreshExpiredToken,
-    policy, healthStore, eventStore, quotaMonitor,
-  };
-
-  const { candidates, preferredAffinity, provisionalAffinity } = await buildFallbackCandidates(deps);
-
-  if (candidates.length === 0) {
-    return emptyResult(0);
-  }
-
-  let preferredReassignmentAllowed = preferredAffinity === undefined || candidates.every(
-    (candidate) => candidate.providerId !== preferredAffinity.providerId || candidate.account.accountId !== preferredAffinity.accountId,
-  );
-  const hasStickyAffinity = Boolean(promptCacheKey && (preferredAffinity || provisionalAffinity));
-  let stickyTransportFailureCandidates = 0;
-  let abortRemainingCandidatesForStickyTransportFailure = false;
-
-  for (const [candidateIndex, candidate] of candidates.entries()) {
-    const candidateStrategy = selectRemoteProviderStrategyForRoute(context, candidate.providerId);
-    let candidatePayload = candidateStrategy === strategy
-      ? payload
-      : candidateStrategy.buildPayload(context);
-
-    // Requesty requires model names in "provider/model" format (e.g., "openai/gpt-5.4").
-    if (candidate.providerId.trim().toLowerCase() === "requesty") {
-      const model = typeof candidatePayload.upstreamPayload.model === "string"
-        ? candidatePayload.upstreamPayload.model
-        : "";
-      if (model && !model.includes("/")) {
-        const prefix = requestyModelPrefix(model);
-        const prefixed = { ...candidatePayload.upstreamPayload, model: `${prefix}/${model}` };
-        candidatePayload = { ...candidatePayload, upstreamPayload: prefixed, bodyText: JSON.stringify(prefixed) };
-      }
-    }
-
-    // Ollama Cloud only accepts "high", "medium", "low", or "none" for reasoning effort.
-    // Normalize "xhigh" -> "high" for this provider.
-    if (candidate.providerId.trim().toLowerCase() === "ollama-cloud") {
-      const normalized = normalizeReasoningEffortForOllamaCloud(candidatePayload.upstreamPayload);
-      if (normalized !== candidatePayload.upstreamPayload) {
-        candidatePayload = { ...candidatePayload, upstreamPayload: normalized, bodyText: JSON.stringify(normalized) };
-      }
-    }
-
-    const hasMoreCandidates = candidateIndex < candidates.length - 1;
-    const releaseInFlight = keyPool.markInFlight(candidate.account);
-
-    for (let retryIndex = 0; retryIndex <= context.config.upstreamTransientRetryCount; retryIndex += 1) {
-      const baseProviderContext: Omit<ProviderAttemptContext, "attempt"> = {
-        ...context,
-        providerId: candidate.providerId,
-        // This may be overridden per-attempt when `OPENAI_IMAGES_UPSTREAM_MODE=platform|auto`.
-        baseUrl: candidate.baseUrl,
-        account: candidate.account,
-        hasMoreCandidates,
-      };
-
-      const primaryUpstreamPath = candidateStrategy.getUpstreamPath(baseProviderContext);
-      const isOpenAiImages = candidate.providerId === context.config.openaiProviderId && candidateStrategy.mode === "images";
-
-      type UpstreamStepKind = "default" | "openai_platform" | "openai_chatgpt" | "openai_codex_responses_images";
-      type UpstreamStep = {
-        readonly baseUrl: string;
-        readonly upstreamPath: string;
-        readonly kind: UpstreamStepKind;
-        /** HTTP statuses that should fall through to the *next base URL* step (used for platform → ChatGPT). */
-        readonly fallbackToNextBaseOnStatuses?: readonly number[];
-      };
-
-      const upstreamSteps: readonly UpstreamStep[] = (() => {
-        if (!isOpenAiImages) {
-          return [{ baseUrl: candidate.baseUrl, upstreamPath: primaryUpstreamPath, kind: "default" as const }];
-        }
-        const mode = context.config.openaiImagesUpstreamMode;
-
-        // API keys should always use the Platform endpoint.
-        if (candidate.account.authType === "api_key") {
-          return [{ baseUrl: context.config.openaiApiBaseUrl, upstreamPath: primaryUpstreamPath, kind: "openai_platform" as const }];
-        }
-
-        // ChatGPT mode uses the Codex backend Responses stream + the built-in `image_generation`
-        // tool, then translates the result back into an Images API-compatible response.
-        if (mode === "chatgpt") {
-          return [{
-            baseUrl: context.config.openaiBaseUrl,
-            upstreamPath: context.config.openaiResponsesPath,
-            kind: "openai_codex_responses_images" as const,
-          }];
-        }
-
-        if (mode === "platform") {
-          return [{ baseUrl: context.config.openaiApiBaseUrl, upstreamPath: primaryUpstreamPath, kind: "openai_platform" as const }];
-        }
-
-        // auto: try Platform Images API first, then fall back to Codex Responses image generation
-        // on auth/scope failures.
-        return [
-          {
-            baseUrl: context.config.openaiApiBaseUrl,
-            upstreamPath: primaryUpstreamPath,
-            kind: "openai_platform" as const,
-            fallbackToNextBaseOnStatuses: [401, 403],
-          },
-          {
-            baseUrl: context.config.openaiBaseUrl,
-            upstreamPath: context.config.openaiResponsesPath,
-            kind: "openai_codex_responses_images" as const,
-          },
-        ];
-      })();
-
-      const hasRetryRemaining = retryIndex < context.config.upstreamTransientRetryCount;
-      let shouldContinueTransientRetry = false;
-
-      for (const [stepIndex, step] of upstreamSteps.entries()) {
-        const upstreamPath = step.upstreamPath;
-        accumulator.attempts += 1;
-        const providerContext: ProviderAttemptContext = {
-          ...baseProviderContext,
-          baseUrl: step.baseUrl,
-          attempt: accumulator.attempts,
-        };
-
-        const upstreamUrl = joinUrl(providerContext.baseUrl, upstreamPath);
-        const upstreamHeaders = buildUpstreamHeadersForCredential(context.clientHeaders, candidate.account, {
-          useOpenAiCodexHeaderProfile: shouldUseOpenAiCodexHeaderProfile(
-            candidate.providerId,
-            candidate.account,
-            context.config.openaiProviderId,
-          ),
-        });
-        candidateStrategy.applyRequestHeaders(upstreamHeaders, providerContext, candidatePayload.upstreamPayload);
-        const attemptStartedAt = Date.now();
-
-        const upstreamSpan = getTelemetry().startSpan("proxy.upstream_attempt", {
-          "proxy.provider_id": candidate.providerId,
-          "proxy.account_id": candidate.account.accountId,
-          "proxy.auth_type": candidate.account.authType,
-          "proxy.upstream_mode": candidateStrategy.mode,
-          "proxy.upstream_path": upstreamPath,
-          "proxy.model": context.routedModel,
-          "proxy.requested_model": context.requestedModelInput,
-          "proxy.base_url": providerContext.baseUrl,
-          "proxy.fallback_attempt": accumulator.attempts,
-        });
-        upstreamSpan.setAttributes({
-          "proxy.service_tier": candidatePayload.serviceTier,
-          "proxy.service_tier_source": candidatePayload.serviceTierSource,
-        });
-
-        const isCodexResponsesImages = step.kind === "openai_codex_responses_images";
-        const effectiveBody = isCodexResponsesImages
-          ? buildCodexResponsesImagesBody(candidatePayload.upstreamPayload)
-          : candidatePayload.bodyText;
-
-        // Emit request event to the data lake before sending upstream.
-        const attemptEntryId = `${candidate.providerId}:${candidate.account.accountId}:${Date.now()}`;
-        if (eventStore) {
-          eventStore.emitRequest(
-            attemptEntryId,
-            candidate.providerId,
-            candidate.account.accountId,
-            context.routedModel,
-            candidatePayload.upstreamPayload,
-            {
-              upstreamMode: candidateStrategy.mode,
-              upstreamPath,
-              upstreamUrl,
-              attempt: accumulator.attempts,
-              requestedModel: context.requestedModelInput,
-              serviceTier: candidatePayload.serviceTier,
-            },
-          );
-        }
-
-        let upstreamResponse: Response;
-        try {
-          upstreamResponse = await fetchWithResponseTimeout(upstreamUrl, {
-            method: "POST",
-            headers: upstreamHeaders,
-            body: effectiveBody
-          }, context.upstreamAttemptTimeoutMs);
-        } catch (error) {
-          const latencyMs = Date.now() - attemptStartedAt;
-          upstreamSpan.setAttribute("proxy.latency_ms", latencyMs);
-          upstreamSpan.setAttribute("proxy.status", 0);
-          upstreamSpan.recordError(error);
-          upstreamSpan.end();
-          accumulator.sawRequestError = true;
-          await providerRoutePheromoneStore.noteFailure(candidate.providerId, context.routedModel);
-          const logEntryId = recordAttempt(requestLogStore, providerContext, {
-            providerId: candidate.providerId,
-            accountId: candidate.account.accountId,
-            authType: candidate.account.authType,
-            upstreamPath,
-            status: 0,
-            latencyMs,
-            serviceTier: candidatePayload.serviceTier,
-            serviceTierSource: candidatePayload.serviceTierSource,
-            factoryDiagnostics: buildFactory4xxDiagnostics(candidatePayload.upstreamPayload, promptCacheKey),
-            error: toErrorMessage(error)
-          }, candidateStrategy.mode);
-
-          if (eventStore) {
-            eventStore.emitError(attemptEntryId, candidate.providerId, candidate.account.accountId, context.routedModel, 0, {
-              error: toErrorMessage(error),
-              logEntryId,
-            }, { latencyMs });
-          }
-          if (hasStickyAffinity) {
-            stickyTransportFailureCandidates += 1;
-            if (stickyTransportFailureCandidates >= MAX_STICKY_TRANSPORT_FAILURE_CANDIDATES) {
-              abortRemainingCandidatesForStickyTransportFailure = true;
-            }
-          }
-          break;
-        }
-
-        const latencyMs = Date.now() - attemptStartedAt;
-        upstreamSpan.setAttribute("proxy.status", upstreamResponse.status);
-        upstreamSpan.setAttribute("proxy.latency_ms", latencyMs);
-
-        const requestLogEntryId = recordAttempt(requestLogStore, providerContext, {
-          providerId: candidate.providerId,
-          accountId: candidate.account.accountId,
-          authType: candidate.account.authType,
-          upstreamPath,
-          status: upstreamResponse.status,
-          latencyMs,
-          serviceTier: candidatePayload.serviceTier,
-          serviceTierSource: candidatePayload.serviceTierSource,
-          factoryDiagnostics: buildFactory4xxDiagnostics(candidatePayload.upstreamPayload, promptCacheKey),
-          promptCacheKeyUsed: Boolean(promptCacheKey),
-        }, candidateStrategy.mode);
-
-        const diagnosticsPromise = updateFailedAttemptDiagnostics(
-          requestLogStore,
-          requestLogEntryId,
-          upstreamResponse,
-          candidate.providerId,
-          candidatePayload.upstreamPayload,
-          promptCacheKey,
-        );
-
-        // Emit error event for non-OK responses with the response body.
-        if (eventStore && !upstreamResponse.ok) {
-          try {
-            const errorBody = await upstreamResponse.clone().json() as Record<string, unknown>;
-            eventStore.emitError(
-              attemptEntryId, candidate.providerId, candidate.account.accountId,
-              context.routedModel, upstreamResponse.status, errorBody,
-              { latencyMs, logEntryId: requestLogEntryId },
-            );
-          } catch {
-            eventStore.emitError(
-              attemptEntryId, candidate.providerId, candidate.account.accountId,
-              context.routedModel, upstreamResponse.status,
-              { status: upstreamResponse.status, statusText: upstreamResponse.statusText },
-              { latencyMs, logEntryId: requestLogEntryId },
-            );
-          }
-        }
-
-        const usagePromise = updateUsageCountsFromResponse(
-          requestLogStore,
-          requestLogEntryId,
-          upstreamResponse,
-          candidateStrategy.mode,
-          context.routedModel,
-          candidate.providerId,
-          context.config,
-          attemptStartedAt,
-        );
-        if (responseLooksLikeEventStream(upstreamResponse, candidateStrategy.mode) && context.clientWantsStream) {
-          void usagePromise;
-        } else {
-          await usagePromise;
-        }
-        await diagnosticsPromise;
-
-        // OpenAI image generation can target either Platform Images API (`api.openai.com`) or the
-        // ChatGPT Codex Responses backend (`chatgpt.com/backend-api/codex/responses`).
-        // When `OPENAI_IMAGES_UPSTREAM_MODE=auto`, fall back from Platform → Codex Responses on
-        // scope/auth failures.
-        if (isOpenAiImages) {
-          const nextStep = stepIndex < upstreamSteps.length - 1 ? upstreamSteps[stepIndex + 1] : undefined;
-          const canFallbackToNextBase =
-            step.fallbackToNextBaseOnStatuses?.includes(upstreamResponse.status) === true
-            && nextStep
-            && nextStep.baseUrl !== step.baseUrl;
-
-          if (canFallbackToNextBase) {
-            try {
-              await upstreamResponse.arrayBuffer();
-            } catch {
-              // ignore
-            }
-            upstreamSpan.setStatus("error", "openai_images_fallback_to_next_base");
-            upstreamSpan.end();
-            continue;
-          }
-
-          // NOTE: We intentionally do not attempt alternate paths for Codex Responses image
-          // generation. If the configured Codex Responses path is invalid, treat it as an upstream
-          // rejection.
-        }
-
-        if (isRateLimitResponse(upstreamResponse)) {
-          accumulator.sawRateLimit = true;
-
-          let ollamaMultiplier = 1;
-          let ollamaLimitKind: ReturnType<typeof detectOllamaLimitKind> = "unknown";
-          let responseBody: Record<string, unknown> | undefined;
-          if (candidate.account.providerId === "ollama-cloud") {
-            try {
-              const cloned = upstreamResponse.clone();
-              responseBody = await cloned.json() as Record<string, unknown>;
-              ollamaLimitKind = detectOllamaLimitKind(responseBody);
-              if (ollamaLimitKind === "weekly") {
-                ollamaMultiplier = context.config.ollamaWeeklyCooldownMultiplier;
-              }
-            } catch {
-              // If we can't parse the body, fall back to no multiplier.
-            }
-          }
-
-          let cooldownMs: number | undefined;
-          if (quotaMonitor?.tracksProvider(candidate.account.providerId)) {
-            cooldownMs = quotaMonitor.getCooldownMs(candidate.account.accountId);
-          }
-          if (!cooldownMs) {
-            cooldownMs = await extractRateLimitCooldownMs(upstreamResponse);
-          }
-          if (!cooldownMs && quotaMonitor?.tracksProvider(candidate.account.providerId)) {
-            try {
-              await quotaMonitor.refreshAccountQuota(candidate.account.accountId);
-            } catch {
-              // Ignore quota lookup failures and fall back to response-derived cooldowns.
-            }
-            cooldownMs = quotaMonitor.getCooldownMsFromQuota(candidate.account.accountId);
-          }
-          if (!cooldownMs) {
-            cooldownMs = context.config.keyCooldownMs;
-          }
-
-          cooldownMs = Math.round(cooldownMs * ollamaMultiplier);
-
-          // Parse the response body once for classification if not already done.
-          if (!responseBody) {
-            try {
-              responseBody = await upstreamResponse.clone().json() as Record<string, unknown>;
-            } catch {
-              responseBody = undefined;
-            }
-          }
-
-          const rateLimitKind = classifyRateLimitKind(
-            responseBody,
-            cooldownMs,
-            context.config.concurrencyThrottleThresholdMs,
-          );
-
-          if (rateLimitKind === "concurrency_throttle") {
-            // Concurrency throttle: wait for the indicated period, then retry
-            // the same credential instead of falling over to a different provider.
-            const maxRetries = context.config.concurrencyThrottleMaxRetries;
-            const waitMs = Math.min(cooldownMs, 10_000);
-            upstreamSpan.setAttribute("proxy.rate_limit_kind", "concurrency_throttle");
-            upstreamSpan.setAttribute("proxy.concurrency_retry_wait_ms", waitMs);
-
-            // Mark a short cooldown so the same account isn't selected by
-            // another concurrent request while we're waiting.
-            keyPool.markRateLimited(candidate.account, Math.min(cooldownMs, 15_000));
-
-            for (let concurrencyRetry = 0; concurrencyRetry < maxRetries; concurrencyRetry += 1) {
-              await sleep(waitMs);
-
-              // Clear the short cooldown so we can re-use this account.
-              if (keyPool.clearAccountCooldown) {
-                keyPool.clearAccountCooldown(candidate.account.providerId, candidate.account.accountId);
-              }
-
-              const retryUrl = joinUrl(providerContext.baseUrl, primaryUpstreamPath);
-              const retryHeaders = buildUpstreamHeadersForCredential(context.clientHeaders, candidate.account, {
-                useOpenAiCodexHeaderProfile: shouldUseOpenAiCodexHeaderProfile(
-                  candidate.providerId,
-                  candidate.account,
-                  context.config.openaiProviderId,
-                ),
-              });
-              candidateStrategy.applyRequestHeaders(retryHeaders, providerContext, candidatePayload.upstreamPayload);
-
-              let retryResponse: Response;
-              try {
-                retryResponse = await fetchWithResponseTimeout(retryUrl, {
-                  method: "POST",
-                  headers: retryHeaders,
-                  body: effectiveBody,
-                }, context.upstreamAttemptTimeoutMs);
-              } catch {
-                // Transport error on retry — fall through to normal fallback.
-                break;
-              }
-
-              if (!isRateLimitResponse(retryResponse)) {
-                // No longer rate-limited: handle the response normally.
-                const retryLatencyMs = Date.now() - attemptStartedAt;
-                const retryLogId = recordAttempt(requestLogStore, providerContext, {
-                  providerId: candidate.providerId,
-                  accountId: candidate.account.accountId,
-                  authType: candidate.account.authType,
-                  upstreamPath,
-                  status: retryResponse.status,
-                  latencyMs: retryLatencyMs,
-                  serviceTier: candidatePayload.serviceTier,
-                  serviceTierSource: candidatePayload.serviceTierSource,
-                  factoryDiagnostics: buildFactory4xxDiagnostics(candidatePayload.upstreamPayload, promptCacheKey),
-                }, candidateStrategy.mode);
-
-                const retryUsagePromise = updateUsageCountsFromResponse(
-                  requestLogStore,
-                  retryLogId,
-                  retryResponse,
-                  candidateStrategy.mode,
-                  context.routedModel,
-                  candidate.providerId,
-                  context.config,
-                  attemptStartedAt,
-                );
-
-                if (retryResponse.ok) {
-                  await retryUsagePromise;
-                  if (healthStore) {
-                    healthStore.recordSuccess(candidate.account, retryResponse.status);
-                  }
-                  if (candidate.account.providerId === "ollama-cloud" && keyPool.clearAccountCooldown) {
-                    keyPool.clearAccountCooldown(candidate.account.providerId, candidate.account.accountId);
-                  }
-                  await providerRoutePheromoneStore.noteSuccess(
-                    candidate.providerId,
-                    context.routedModel,
-                    clampRouteQuality(retryLatencyMs),
-                  );
-                  upstreamSpan.setStatus("ok");
-                  upstreamSpan.end();
-                  releaseInFlight();
-                  return { handled: true, candidateCount: candidates.length, summary: accumulator };
-                }
-
-                // Non-429 error on retry — accumulate and break to fallback loop.
-                accumulator.sawUpstreamServerError ||= retryResponse.status >= 500;
-                accumulator.sawUpstreamInvalidRequest ||= retryResponse.status >= 400 && retryResponse.status < 500;
-                try { await retryResponse.arrayBuffer(); } catch { /* ignore */ }
-                break;
-              }
-
-              // Still 429 on retry — re-classify and either wait again or give up.
-              let retryCooldownMs = await extractRateLimitCooldownMs(retryResponse);
-              if (!retryCooldownMs) {
-                retryCooldownMs = cooldownMs;
-              }
-              let retryBody: Record<string, unknown> | undefined;
-              try {
-                retryBody = await retryResponse.clone().json() as Record<string, unknown>;
-              } catch {
-                retryBody = undefined;
-              }
-              const retryKind = classifyRateLimitKind(retryBody, retryCooldownMs, context.config.concurrencyThrottleThresholdMs);
-              try { await retryResponse.arrayBuffer(); } catch { /* ignore */ }
-
-              if (retryKind !== "concurrency_throttle") {
-                // Escalated to quota exhaustion — stop retrying this credential.
-                keyPool.markRateLimited(candidate.account, Math.round(retryCooldownMs * ollamaMultiplier));
-                break;
-              }
-
-              // Still a concurrency throttle — mark short cooldown and loop again.
-              keyPool.markRateLimited(candidate.account, Math.min(retryCooldownMs, 15_000));
-            }
-
-            // All concurrency retries exhausted — mark normal cooldown and fall over.
-            keyPool.markRateLimited(candidate.account, cooldownMs);
-          } else {
-            // Quota exhaustion: mark rate-limited and fall over to next candidate.
-            keyPool.markRateLimited(candidate.account, cooldownMs);
-          }
-
-          if (
-            ollamaLimitKind === "session"
-            && promptCacheKey
-            && (
-              candidateMatchesAffinity(candidate, preferredAffinity)
-              || candidateMatchesAffinity(candidate, provisionalAffinity)
-            )
-          ) {
-            await promptAffinityStore.delete(promptCacheKey);
-            preferredReassignmentAllowed = true;
-          }
-          if (
-            preferredAffinity
-            && candidate.providerId === preferredAffinity.providerId
-            && candidate.account.accountId === preferredAffinity.accountId
-          ) {
-            preferredReassignmentAllowed = true;
-          }
-          upstreamSpan.setStatus("error", rateLimitKind === "concurrency_throttle" ? "concurrency_throttle" : "rate_limited");
-          upstreamSpan.end();
-          break;
-        }
-
-        if (await responseIndicatesQuotaError(upstreamResponse)) {
-          accumulator.sawRateLimit = true;
-          const permanentlyDisable = shouldPermanentlyDisableCredential(candidate.account, upstreamResponse.status);
-          const baseCooldownMs = permanentlyDisable
-            ? PERMANENT_DISABLE_COOLDOWN_MS
-            : upstreamResponse.status === 402
-              ? 24 * 60 * 60 * 1000
-              : Math.min(context.config.keyCooldownMs, 60_000);
-          let quotaCooldownMs: number | undefined;
-          if (quotaMonitor?.tracksProvider(candidate.account.providerId)) {
-            quotaCooldownMs = quotaMonitor.getCooldownMs(candidate.account.accountId);
-            if (!quotaCooldownMs) {
-              try {
-                await quotaMonitor.refreshAccountQuota(candidate.account.accountId);
-              } catch {
-                // Ignore quota lookup failures and fall back to local cooldown heuristics.
-              }
-              quotaCooldownMs = quotaMonitor.getCooldownMs(candidate.account.accountId);
-            }
-          }
-          if (!quotaCooldownMs) {
-            quotaCooldownMs = await extractRateLimitCooldownMs(upstreamResponse);
-          }
-          if (!quotaCooldownMs) {
-            quotaCooldownMs = healthStore
-              ? healthStore.getGrowingCooldown(candidate.account.providerId, candidate.account.accountId, baseCooldownMs)
-              : baseCooldownMs;
-          }
-          keyPool.markRateLimited(candidate.account, quotaCooldownMs);
-          if (healthStore) {
-            healthStore.recordFailure(candidate.account, upstreamResponse.status, "quota_exhausted");
-          }
-          await providerRoutePheromoneStore.noteFailure(candidate.providerId, context.routedModel);
-          if (
-            preferredAffinity
-            && candidate.providerId === preferredAffinity.providerId
-            && candidate.account.accountId === preferredAffinity.accountId
-          ) {
-            preferredReassignmentAllowed = true;
-          }
-          try {
-            await upstreamResponse.arrayBuffer();
-          } catch {
-            // Ignore body read failures while failing over.
-          }
-          upstreamSpan.setStatus("error", "quota_exhausted");
-          upstreamSpan.end();
-          break;
-        }
-
-        // Factory intermittently returns 403 during auth token rotation; retry once
-        // just like a server error so the next attempt often succeeds.
-        if (upstreamResponse.status === 403 && candidate.providerId === "factory" && hasRetryRemaining) {
-          try { await upstreamResponse.arrayBuffer(); } catch { /* ignore */ }
-          upstreamSpan.setStatus("error", "factory_transient_403");
-          upstreamSpan.end();
-          await sleep(transientRetryDelayMs(context, retryIndex));
-          shouldContinueTransientRetry = true;
-          break;
-        }
-
-        if (upstreamResponse.status >= 500 && upstreamResponse.status <= 599) {
-          accumulator.sawUpstreamServerError = true;
-          if (hasRetryRemaining && shouldRetrySameCredentialForServerError(upstreamResponse.status)) {
-            try {
-              await upstreamResponse.arrayBuffer();
-            } catch {
-              // Ignore body read failures while retrying.
-            }
-            upstreamSpan.setStatus("error", `upstream_server_error_${upstreamResponse.status}`);
-            upstreamSpan.end();
-            await sleep(transientRetryDelayMs(context, retryIndex));
-            shouldContinueTransientRetry = true;
-            break;
-          }
-          keyPool.markRateLimited(candidate.account, Math.min(context.config.keyCooldownMs, 5000));
-          await providerRoutePheromoneStore.noteFailure(candidate.providerId, context.routedModel);
-          try {
-            await upstreamResponse.arrayBuffer();
-          } catch {
-            // Ignore body read failures while failing over.
-          }
-          upstreamSpan.setStatus("error", `upstream_server_error_${upstreamResponse.status}`);
-          upstreamSpan.end();
-          break;
-        }
-
-        // Codex Responses API → Images API translation: read the streaming Responses output,
-        // extract `image_generation_call` results, and synthesize an Images API JSON response.
-        if (isCodexResponsesImages && upstreamResponse.ok) {
-          const responseText = await upstreamResponse.text();
-          const contentType = upstreamResponse.headers.get("content-type") ?? "";
-          const looksLikeEventStream = contentType.toLowerCase().includes("text/event-stream") || contentType.length === 0;
-
-          // Check for errors in the SSE stream.
-          if (looksLikeEventStream) {
-            const streamError = responsesEventStreamToErrorPayload(responseText);
-            if (streamError) {
-              upstreamSpan.setStatus("error", "codex_responses_stream_error");
-              upstreamSpan.end();
-              accumulator.sawUpstreamInvalidRequest = true;
-              break;
-            }
-          }
-
-          const imagesPayload = looksLikeEventStream
-            ? extractImagesFromCodexEventStream(responseText)
-            : extractImagesFromCodexResponse(responseText);
-
-          if (imagesPayload) {
-            reply.header("x-open-hax-upstream-provider", providerContext.providerId);
-            reply.header("x-open-hax-upstream-mode", "codex_responses_images");
-            reply.code(200);
-            reply.header("content-type", "application/json");
-            reply.send(imagesPayload);
-            upstreamSpan.setStatus("ok");
-            upstreamSpan.end();
-            if (healthStore) {
-              healthStore.recordSuccess(candidate.account, upstreamResponse.status);
-            }
-            if (candidate.account.providerId === "ollama-cloud" && keyPool.clearAccountCooldown) {
-              keyPool.clearAccountCooldown(candidate.account.providerId, candidate.account.accountId);
-            }
-            await providerRoutePheromoneStore.noteSuccess(
-              candidate.providerId,
-              context.routedModel,
-              clampRouteQuality(latencyMs),
-            );
-            releaseInFlight();
-            return { handled: true, candidateCount: candidates.length, summary: accumulator };
-          }
-
-          // Responses completed but contained no image_generation_call outputs.
-          upstreamSpan.setStatus("error", "codex_responses_no_image_output");
-          upstreamSpan.end();
-          accumulator.sawUpstreamInvalidRequest = true;
-          break;
-        }
-
-        reply.header("x-open-hax-upstream-mode", candidateStrategy.mode);
-        const outcome = await candidateStrategy.handleProviderAttempt(reply, upstreamResponse, providerContext);
-        if (outcome.kind === "handled") {
-          upstreamSpan.setStatus("ok");
-          upstreamSpan.end();
-          if (healthStore && upstreamResponse.ok) {
-            healthStore.recordSuccess(candidate.account, upstreamResponse.status);
-          }
-          if (upstreamResponse.ok && candidate.account.providerId === "ollama-cloud" && keyPool.clearAccountCooldown) {
-            keyPool.clearAccountCooldown(candidate.account.providerId, candidate.account.accountId);
-          }
-          await providerRoutePheromoneStore.noteSuccess(
-            candidate.providerId,
-            context.routedModel,
-            clampRouteQuality(latencyMs),
-          );
-          if (eventStore) {
-            eventStore.emitResponse(
-              attemptEntryId, candidate.providerId, candidate.account.accountId,
-              context.routedModel, upstreamResponse.status, null,
-              { latencyMs: Date.now() - attemptStartedAt, logEntryId: requestLogEntryId },
-            );
-          }
-          if (
-            promptCacheKey
-            && (
-              preferredAffinity === undefined
-              || preferredReassignmentAllowed
-              || (candidate.providerId === preferredAffinity.providerId && candidate.account.accountId === preferredAffinity.accountId)
-            )
-          ) {
-            await promptAffinityStore.noteSuccess(promptCacheKey, candidate.providerId, candidate.account.accountId);
-          }
-          releaseInFlight();
-          return {
-            handled: true,
-            candidateCount: candidates.length,
-            summary: accumulator
-          };
-        }
-
-        if (
-          healthStore
-          && !upstreamResponse.ok
-          && upstreamResponse.status >= 500
-          && !outcome.upstreamInvalidRequest
-          && !outcome.modelNotFound
-          && !outcome.modelNotSupportedForAccount
-        ) {
-          healthStore.recordFailure(candidate.account, upstreamResponse.status);
-        }
-        if (!upstreamResponse.ok) {
-          await providerRoutePheromoneStore.noteFailure(candidate.providerId, context.routedModel);
-          if (
-            preferredAffinity
-            && candidate.providerId === preferredAffinity.providerId
-            && candidate.account.accountId === preferredAffinity.accountId
-          ) {
-            preferredReassignmentAllowed = true;
-          }
-        }
-
-        /**
-         * Handle hidden upstream errors: providers that return 200 OK but carry a
-         * quota error ("stream_quota_error") or empty/invalid body ("stream_empty_or_invalid")
-         * in the SSE stream. These accounts must be put into cooldown and the sticky
-         * affinity record deleted so the fallback loop can try the next candidate.
-         */
-        if (upstreamResponse.ok && (outcome.rateLimit === true || outcome.requestError === true)) {
-          const cooldownMs = outcome.rateLimit === true
-            ? context.config.keyCooldownMs
-            : Math.min(context.config.keyCooldownMs, 10_000);
-          keyPool.markRateLimited(candidate.account, cooldownMs);
-          await providerRoutePheromoneStore.noteFailure(candidate.providerId, context.routedModel);
-          if (outcome.rateLimit === true) {
-            requestLogStore.update(requestLogEntryId, {
-              upstreamErrorCode: "stream_quota_error",
-              upstreamErrorMessage: "200 OK with quota error in stream body",
-            });
-          } else {
-            requestLogStore.update(requestLogEntryId, {
-              upstreamErrorCode: "stream_empty_or_invalid",
-              upstreamErrorMessage: "200 OK with no substantive content in stream body",
-            });
-          }
-          if (
-            candidateMatchesAffinity(candidate, preferredAffinity)
-            || candidateMatchesAffinity(candidate, provisionalAffinity)
-          ) {
-            preferredReassignmentAllowed = true;
-          }
-          if (outcome.rateLimit === true && promptCacheKey && (candidateMatchesAffinity(candidate, preferredAffinity) || candidateMatchesAffinity(candidate, provisionalAffinity))) {
-            await promptAffinityStore.delete(promptCacheKey);
-          }
-        }
-
-        accumulator.sawRateLimit ||= outcome.rateLimit === true;
-        accumulator.sawRequestError ||= outcome.requestError === true;
-        accumulator.sawUpstreamServerError ||= outcome.upstreamServerError === true;
-        accumulator.sawUpstreamInvalidRequest ||= outcome.upstreamInvalidRequest === true;
-        accumulator.sawModelNotFound ||= outcome.modelNotFound === true;
-        accumulator.sawModelNotSupportedForAccount ||= outcome.modelNotSupportedForAccount === true;
-        if (outcome.upstreamAuthError) {
-          accumulator.lastUpstreamAuthError = outcome.upstreamAuthError;
-        }
-
-        if (!upstreamResponse.ok && outcome.requestError === true && upstreamResponse.status === 401 && candidate.account.authType === "oauth_bearer" && candidate.account.refreshToken && refreshExpiredToken) {
-          const refreshedCredential = await refreshExpiredToken(candidate.account);
-          if (refreshedCredential) {
-            const refreshedProviderContext: ProviderAttemptContext = { ...providerContext, account: refreshedCredential };
-            const refreshedHeaders = buildUpstreamHeadersForCredential(context.clientHeaders, refreshedCredential, {
-              useOpenAiCodexHeaderProfile: shouldUseOpenAiCodexHeaderProfile(
-                candidate.providerId,
-                refreshedCredential,
-                context.config.openaiProviderId,
-              ),
-            });
-            candidateStrategy.applyRequestHeaders(refreshedHeaders, refreshedProviderContext, candidatePayload.upstreamPayload);
-            const refreshedRelease = keyPool.markInFlight(refreshedCredential);
-            const refreshedAttemptStartedAt = Date.now();
-            let refreshedResponse: Response;
-            try {
-              refreshedResponse = await fetchWithResponseTimeout(upstreamUrl, {
-                method: "POST",
-                headers: refreshedHeaders,
-                body: effectiveBody
-              }, context.upstreamAttemptTimeoutMs);
-            } catch (error) {
-              refreshedRelease();
-              releaseInFlight();
-              throw error;
-            }
-
-            try {
-              const refreshedLatencyMs = Date.now() - refreshedAttemptStartedAt;
-              const refreshedLogId = recordAttempt(requestLogStore, refreshedProviderContext, {
-                providerId: candidate.providerId,
-                accountId: refreshedCredential.accountId,
-                authType: refreshedCredential.authType,
-                upstreamPath,
-                status: refreshedResponse.status,
-                latencyMs: refreshedLatencyMs,
-                serviceTier: candidatePayload.serviceTier,
-                serviceTierSource: candidatePayload.serviceTierSource,
-                factoryDiagnostics: buildFactory4xxDiagnostics(candidatePayload.upstreamPayload, promptCacheKey),
-                promptCacheKeyUsed: Boolean(promptCacheKey),
-              }, candidateStrategy.mode);
-              const refreshedDiagnosticsPromise = updateFailedAttemptDiagnostics(
-                requestLogStore,
-                refreshedLogId,
-                refreshedResponse,
-                candidate.providerId,
-                candidatePayload.upstreamPayload,
-                promptCacheKey,
-              );
-              const usagePromise = updateUsageCountsFromResponse(
-                requestLogStore,
-                refreshedLogId,
-                refreshedResponse,
-                candidateStrategy.mode,
-                context.routedModel,
-                candidate.providerId,
-                context.config,
-                refreshedAttemptStartedAt,
-              );
-              if (responseLooksLikeEventStream(refreshedResponse, candidateStrategy.mode) && context.clientWantsStream) {
-                void usagePromise;
-              } else {
-                await usagePromise;
-              }
-              await refreshedDiagnosticsPromise;
-              if (isRateLimitResponse(refreshedResponse)) {
-                accumulator.sawRateLimit = true;
-                const refreshedCooldownMs = await extractRateLimitCooldownMs(refreshedResponse);
-                keyPool.markRateLimited(refreshedCredential, refreshedCooldownMs);
-                try {
-                  await refreshedResponse.arrayBuffer();
-                } catch {
-                  // Ignore body read failures while failing over after refresh.
-                }
-                break;
-              }
-              // Handle Codex Responses → Images translation for the refreshed response.
-              if (isCodexResponsesImages && refreshedResponse.ok) {
-                const refreshedText = await refreshedResponse.text();
-                const refreshedContentType = refreshedResponse.headers.get("content-type") ?? "";
-                const refreshedLooksLikeEventStream = refreshedContentType.toLowerCase().includes("text/event-stream") || refreshedContentType.length === 0;
-
-                if (refreshedLooksLikeEventStream) {
-                  const streamError = responsesEventStreamToErrorPayload(refreshedText);
-                  if (streamError) {
-                    upstreamSpan.setStatus("error", "codex_responses_stream_error");
-                    upstreamSpan.end();
-                    accumulator.sawUpstreamInvalidRequest = true;
-                    break;
-                  }
-                }
-
-                const refreshedImagesPayload = refreshedLooksLikeEventStream
-                  ? extractImagesFromCodexEventStream(refreshedText)
-                  : extractImagesFromCodexResponse(refreshedText);
-
-                if (refreshedImagesPayload) {
-                  reply.header("x-open-hax-upstream-provider", providerContext.providerId);
-                  reply.header("x-open-hax-upstream-mode", "codex_responses_images");
-                  reply.code(200);
-                  reply.header("content-type", "application/json");
-                  reply.send(refreshedImagesPayload);
-                  upstreamSpan.setStatus("ok");
-                  upstreamSpan.end();
-                  if (healthStore) {
-                    healthStore.recordSuccess(refreshedCredential, refreshedResponse.status);
-                  }
-                  await providerRoutePheromoneStore.noteSuccess(
-                    candidate.providerId,
-                    context.routedModel,
-                    clampRouteQuality(refreshedLatencyMs),
-                  );
-                  releaseInFlight();
-                  return { handled: true, candidateCount: candidates.length, summary: accumulator };
-                }
-                upstreamSpan.setStatus("error", "codex_responses_no_image_output");
-                upstreamSpan.end();
-                accumulator.sawUpstreamInvalidRequest = true;
-                break;
-              }
-
-              reply.header("x-open-hax-upstream-mode", candidateStrategy.mode);
-              const refreshedOutcome = await candidateStrategy.handleProviderAttempt(reply, refreshedResponse, refreshedProviderContext);
-              if (refreshedOutcome.kind === "handled") {
-                upstreamSpan.setStatus("ok");
-                upstreamSpan.end();
-
-                if (healthStore && refreshedResponse.ok) {
-                  healthStore.recordSuccess(refreshedCredential, refreshedResponse.status);
-                }
-                await providerRoutePheromoneStore.noteSuccess(
-                  candidate.providerId,
-                  context.routedModel,
-                  clampRouteQuality(refreshedLatencyMs),
-                );
-
-                if (
-                  promptCacheKey
-                  && (
-                    preferredAffinity === undefined
-                    || preferredReassignmentAllowed
-                    || (candidate.providerId === preferredAffinity.providerId && refreshedCredential.accountId === preferredAffinity.accountId)
-                  )
-                ) {
-                  await promptAffinityStore.noteSuccess(promptCacheKey, candidate.providerId, refreshedCredential.accountId);
-                }
-
-                releaseInFlight();
-                return { handled: true, candidateCount: candidates.length, summary: accumulator };
-              }
-              accumulator.sawRateLimit ||= refreshedOutcome.rateLimit === true;
-              accumulator.sawRequestError ||= refreshedOutcome.requestError === true;
-              accumulator.sawUpstreamServerError ||= refreshedOutcome.upstreamServerError === true;
-              accumulator.sawUpstreamInvalidRequest ||= refreshedOutcome.upstreamInvalidRequest === true;
-              accumulator.sawModelNotFound ||= refreshedOutcome.modelNotFound === true;
-              accumulator.sawModelNotSupportedForAccount ||= refreshedOutcome.modelNotSupportedForAccount === true;
-              if (!refreshedResponse.ok) {
-                await providerRoutePheromoneStore.noteFailure(candidate.providerId, context.routedModel);
-              }
-              if (refreshedOutcome.upstreamAuthError) {
-                accumulator.lastUpstreamAuthError = refreshedOutcome.upstreamAuthError;
-              }
-              if (!refreshedResponse.ok && refreshedOutcome.requestError === true && (refreshedResponse.status === 401 || refreshedResponse.status === 403)) {
-                if (shouldCooldownCredentialOnAuthFailure(candidate.providerId, refreshedResponse.status)) {
-                  keyPool.markRateLimited(refreshedCredential, Math.min(context.config.keyCooldownMs, 10_000));
-                  // Disable OAuth accounts that fail auth even after successful token refresh
-                  if (refreshedCredential.authType === "oauth_bearer" && keyPool.disableAccount) {
-                    keyPool.disableAccount(refreshedCredential.providerId, refreshedCredential.accountId);
-                  }
-                  if (preferredAffinity && candidate.providerId === preferredAffinity.providerId && refreshedCredential.accountId === preferredAffinity.accountId) {
-                    preferredReassignmentAllowed = true;
-                  }
-                }
-              }
-              break;
-            } finally {
-              refreshedRelease();
-            }
-          } else {
-            await providerRoutePheromoneStore.noteFailure(candidate.providerId, context.routedModel);
-            keyPool.markRateLimited(candidate.account, Math.min(context.config.keyCooldownMs, 10_000));
-            // Disable OAuth accounts when token refresh fails
-            if (candidate.account.authType === "oauth_bearer" && keyPool.disableAccount) {
-              keyPool.disableAccount(candidate.account.providerId, candidate.account.accountId);
-            }
-            if (preferredAffinity && candidate.providerId === preferredAffinity.providerId && candidate.account.accountId === preferredAffinity.accountId) {
-              preferredReassignmentAllowed = true;
-            }
-          }
-        } else if (!upstreamResponse.ok && outcome.requestError === true && (upstreamResponse.status === 401 || upstreamResponse.status === 402 || upstreamResponse.status === 403)) {
-          if (shouldCooldownCredentialOnAuthFailure(candidate.providerId, upstreamResponse.status) || shouldPermanentlyDisableCredential(candidate.account, upstreamResponse.status)) {
-            const permanentlyDisable = shouldPermanentlyDisableCredential(candidate.account, upstreamResponse.status);
-            const cooldownMs = permanentlyDisable
-              ? PERMANENT_DISABLE_COOLDOWN_MS
-              : Math.min(context.config.keyCooldownMs, 10_000);
-            keyPool.markRateLimited(candidate.account, cooldownMs);
-            if (permanentlyDisable && keyPool.disableAccount) {
-              keyPool.disableAccount(candidate.account.providerId, candidate.account.accountId);
-            }
-            // Also disable OAuth accounts with 401 that have no refresh token (unrecoverable)
-            if (upstreamResponse.status === 401 && candidate.account.authType === "oauth_bearer" && !candidate.account.refreshToken && keyPool.disableAccount) {
-              keyPool.disableAccount(candidate.account.providerId, candidate.account.accountId);
-            }
-            if (healthStore) {
-              healthStore.recordFailure(candidate.account, upstreamResponse.status, "credential_disabled");
-            }
-            if (preferredAffinity && candidate.providerId === preferredAffinity.providerId && candidate.account.accountId === preferredAffinity.accountId) {
-              preferredReassignmentAllowed = true;
-            }
-          }
-        }
-
-        // Cooldown accounts that reject the model (e.g. free-tier ChatGPT accounts
-        // that cannot use gpt-5.4). Without this, the same accounts are retried on
-        // every request, causing long cascading failures before reaching a working
-        // provider.
-        if (outcome.modelNotSupportedForAccount === true) {
-          keyPool.markRateLimited(candidate.account, Math.min(context.config.keyCooldownMs, 60_000));
-        }
-
-        if (!upstreamResponse.ok && outcome.requestError === true && !outcome.modelNotFound && !outcome.modelNotSupportedForAccount) {
-          await summarizeUpstreamError(upstreamResponse);
-        }
-
-        upstreamSpan.setStatus("error", `fallback_continue_${upstreamResponse.status}`);
-        upstreamSpan.end();
-        break;
-      }
-
-      if (shouldContinueTransientRetry) {
-        continue;
-      }
-
-      break;
-    }
-
-    releaseInFlight();
-
-    if (abortRemainingCandidatesForStickyTransportFailure) {
-      break;
-    }
-  }
-
-  return {
-    handled: false,
-    candidateCount: candidates.length,
-    summary: accumulator,
-  };
-}
-
-export const executeProviderFallback = executeProviderRoutingPlan;
-
-export async function inspectProviderAvailability(
-  keyPool: {
-    getStatus(providerId: string): Promise<{ readonly totalAccounts: number; readonly disabledAccounts?: number }>;
-  },
-  providerRoutes: readonly ProviderRoute[],
-  promptCacheKey?: string,
-): Promise<ProviderAvailabilitySummary> {
-  let sawConfiguredProvider = false;
-  let sawEnabledConfiguredProvider = false;
-
-  for (const route of providerRoutes) {
-    try {
-      const status = await keyPool.getStatus(route.providerId);
-      if (status.totalAccounts > 0) {
-        sawConfiguredProvider = true;
-        if (status.totalAccounts > (status.disabledAccounts ?? 0)) {
-          sawEnabledConfiguredProvider = true;
-        }
-      }
-    } catch {
-      // Ignore status lookup errors and continue collecting provider info.
-    }
-  }
-
-  return {
-    sawConfiguredProvider,
-    sawOnlyDisabledProviders: sawConfiguredProvider && !sawEnabledConfiguredProvider,
-    prompt_cache_key: promptCacheKey,
-  };
-}
+// Compatibility shim. The execution loop was extracted from `fallback/legacy.ts`
+// into `provider-strategy/routing/attempt-executor.ts`.
+export * from "../routing/attempt-executor.js";
